+++ packages/frontend/src/components/MainSection/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./MainSection";
