+++ packages/frontend/src/components/MessagesSection/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./MessageSection";
