+++ packages/frontend/src/components/PendingChats/PendingChats.tsx
@@ -4,54 +4,15 @@ import { AiFillCheckCircle, AiFillCloseCircle } from "react-icons/ai";
 import { Socket } from "socket.io-client";
 import { requestUrl } from "../../utils/hostUrl_requestUrl";
 import { Iinvites } from "../../pages/[acc]";
-import { FaUserCircle } from "react-icons/fa";
-import styled from "@emotion/styled";
-import { css } from "@emotion/css";
-import Image from "next/image";
-
+import { css, cx } from "@emotion/css";
+import Single_avatar from "../Avatar/Single_avatar";
 interface IPendingChats extends Iinvites {
   _id: string;
   inviter: string;
   status: string;
   reciever: string;
   socketRef: Socket;
-  setLocalStatus: React.Dispatch<React.SetStateAction<string>>;
 }
-
-const Contacts = styled.div`
-  border-top: 2px solid rgba(0, 0, 0, 0.3);
-  background: var(--main-white);
-  width: 95%;
-  margin: 0.5rem;
-  display: flex;
-  justify-content: space-between;
-`;
-
-const Accept = styled.button`
-  color: var(--main-black);
-  outline: none;
-  border: none;
-  font-size: 3.01rem;
-  background: transparent;
-  transition: 0.2s;
-  &:hover {
-    color: var(--main-green);
-    transition: 0.2s;
-  }
-`;
-
-const Decline = styled.button`
-  color: var(--main-black);
-  outline: none;
-  border: none;
-  font-size: 3.01rem;
-  background: transparent;
-  transition: 0.2s;
-  &:hover {
-    color: var(--main-red);
-    transition: 0.2s;
-  }
-`;
 
 const ButtonSharedStyle = `
   display: flex;
@@ -65,37 +26,18 @@ const ButtonSharedStyle = `
     transition: 0.2s;
   }`;
 
-function PendingChats({ _id, inviter, reciever, status, socketRef, setLocalStatus }: IPendingChats) {
-  const [image, setImage] = React.useState<string>("");
+function PendingChats({ _id, inviter, reciever, status, socketRef }: IPendingChats) {
   const emitFriendRequest = async () => {
     socketRef?.emit("friend_request");
   };
 
-  const getUserImage = async (name: string) => {
-    try {
-      const res = await axios.get(`${requestUrl}/users/${name}`);
-      const userAvatar = res.data.user.userAvatar;
-      if (!userAvatar) {
-        setImage("");
-        return false;
-      }
-      const requestString = `${requestUrl}/${userAvatar}`;
-      setImage(requestString);
-      return true;
-    } catch (error) {
-      return false;
-    }
-  };
-
   const updateInviteStatus = async (param: string) => {
     try {
-      setLocalStatus(param);
       await axios.put(`${requestUrl}/invites`, {
         id: _id,
         status: param,
       });
       emitFriendRequest();
-      setLocalStatus("");
 
       return true;
     } catch (error) {
@@ -105,62 +47,97 @@ function PendingChats({ _id, inviter, reciever, status, socketRef, setLocalStatu
 
   const createChatRoom = async () => {
     try {
-      setLocalStatus("accepted");
       await axios.put(`${requestUrl}/chat-room`, {
         id: _id,
         user1: inviter,
         user2: reciever,
       });
       emitFriendRequest();
-      setLocalStatus("");
     } catch (error) {
       return false;
     }
   };
 
-  React.useEffect(() => {
-    getUserImage(inviter);
-  }, []);
-
   return (
     <div style={{ width: "100%" }}>
       {status === "recieved" && (
-        <Contacts className="contacts">
+        <div
+          className={cx(
+            "contacts",
+            css`
+              border-top: 2px solid rgba(0, 0, 0, 0.3);
+              background: var(--main-white);
+              width: 98.5%;
+              height: 20vh;
+              margin: 0.5rem;
+              display: flex;
+              justify-content: space-between;
+            `,
+          )}
+        >
           <div className="user_info flex">
-            {image ? <Image src={image} alt="user_avatar" /> : <FaUserCircle />}
+            <Single_avatar inviter={inviter} cookieName="" width="3rem" height="3rem" />
             <h1 className="flex">{inviter}</h1>
           </div>
           <div className="invite_buttons">
-            <Accept
+            <button
               onClick={() => {
                 createChatRoom();
               }}
-              className={css`
-                ${ButtonSharedStyle}
-              `}
+              className={cx(
+                css`
+                  ${ButtonSharedStyle}
+                `,
+                css`
+                  color: var(--main-black);
+                  outline: none;
+                  border: none;
+                  font-size: 3.01rem;
+                  background: transparent;
+                  transition: 0.2s;
+                  &:hover {
+                    color: var(--main-green);
+                    transition: 0.2s;
+                  }
+                `,
+              )}
             >
               <AiFillCheckCircle
                 className={css`
-                  width: 2rem;
-                  height: 2rem;
+                  width: 3rem;
+                  height: 3rem;
                 `}
               />
-            </Accept>
-            <Decline
+            </button>
+            <button
               onClick={() => updateInviteStatus("declined")}
-              className={css`
-                ${ButtonSharedStyle}
-              `}
+              className={cx(
+                css`
+                  ${ButtonSharedStyle}
+                `,
+                css`
+                  color: var(--main-black);
+                  outline: none;
+                  border: none;
+                  font-size: 3.01rem;
+                  background: transparent;
+                  transition: 0.2s;
+                  &:hover {
+                    color: var(--main-red);
+                    transition: 0.2s;
+                  }
+                `,
+              )}
             >
               <AiFillCloseCircle
                 className={css`
-                  width: 2rem;
-                  height: 2rem;
+                  width: 3rem;
+                  height: 3rem;
                 `}
               />
-            </Decline>
+            </button>
           </div>
-        </Contacts>
+        </div>
       )}
     </div>
   );
