+++ packages/frontend/src/components/RegisterForm/QuickLogin_Modal/QuickLogin_Modal.tsx
@@ -3,8 +3,10 @@ import { Button } from "../../styledComponents";
 import { css, cx } from "@emotion/css";
 import Link from "next/link";
 import Loading from "../../Loading";
+import { useDispatch } from "react-redux";
 
 const QuickLogin_Modal = ({ quickLogin, isLogging }: { quickLogin: () => void; isLogging: boolean }) => {
+  const dispatch = useDispatch();
   return (
     <div style={{ width: "75%", height: "30vh" }} className="fRequests_modal flex">
       <div
@@ -15,31 +17,20 @@ const QuickLogin_Modal = ({ quickLogin, isLogging }: { quickLogin: () => void; i
           "flex",
         )}
       >
-        {isLogging ? (
-          <div
-            className={cx(
-              css`
-                width: 50%;
-              `,
-              "flex",
-            )}
-          >
-            <Loading />
-          </div>
-        ) : (
-          <Button
-            className={css`
-              width: 50%;
-            `}
-            onClick={() => {
-              quickLogin();
-            }}
-          >
-            Click me to Quick login
-          </Button>
-        )}
+        {isLogging && <Loading />}
+        <Button
+          className={css`
+            width: 50%;
+          `}
+          onClick={() => {
+            quickLogin();
+          }}
+        >
+          Click me to Quick login
+        </Button>
         <Link href="/" passHref>
           <Button
+            onClick={() => dispatch({ type: "QUICK_LOGIN", payload: false })}
             className={css`
               width: 50%;
             `}
